// packages/server/src/models/enums.ts
enum ActionType {
  // Feed
  POST_UPDATED = 'POST_UPDATED',
  POST_CREATED = 'POST_CREATED',
  POST_ARCHIVED = 'POST_ARCHIVED',
  POST_REMOVED = 'POST_REMOVED',
  // Event
  EVENT_UPDATED = 'EVENT_UPDATED',
  EVENT_CREATED = 'EVENT_CREATED',
  EVENT_ARCHIVED = 'EVENT_ARCHIVED',
  EVENT_REMOVED = 'EVENT_REMOVED',
  // Network
  NETWORK_UPDATED = 'NETWORK_UPDATED',
  NETWORK_CREATED = 'NETWORK_CREATED',
  NETWORK_ARCHIVED = 'NETWORK_ARCHIVED',
  NETWORK_REMOVED = 'NETWORK_REMOVED',
}

enum EntityType {
  POST = 'Post',
  NETWORK = 'Network',
  EVENT = 'Event',
}

export { ActionType, EntityType };